feat: init project with tsx
